requirements.txt
tensorflow
torch
torchvision
ftfy
regex
tqdm
pillow
numpy
scikit-learn
opencv-python
git+https://github.com/openai/CLIP.git